requirements.txt
asgiref==3.8.1
Django==4.2
django-braces==1.15.0
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-model-utils==4.5.1
django-star-ratings==0.9.2
djangorestframework==3.15.1
pillow==10.3.0
sqlparse==0.5.0
swapper==1.3.0
